retire the 1-felt non-revocation rail (felt-width #11 fold-in, part 2)
REDUNDANCY CONFIRMED BEFORE DELETING (the task's stop-gate): the limb-26
`nullifierFreshOp` `.absent` map-op is the freshness authority over EXACTLY the
set the rail's `DslRevocationTree::revoked_leaves` seeded —
  * same set: `full_turn_proof.rs` threaded `revoked_leaves()` as the limb-26
    BEFORE tree (`before_nullifiers`), node ditto (`turn_proving.rs`);
  * same item: the `.absent` key IS the spend row's folded nullifier `param0`,
    cross-bound to the published ROT_NULLIFIER_PI (the C4 weld) — the rail's
    host binding (b) is structural in-circuit;
  * same root authority, WIDER: the BEFORE limb-26 root8 (~124-bit) is absorbed
    into the OLD commit the verifier pins to the canonical anchor — subsumes the
    rail's 1-felt binding (a);
  * enforcement: forge-rejection canary
    `circuit/tests/vk_epoch_notes_light_client_binding.rs`
    (`notespend_forced_on_wire_rejects_forged_nullifier_root_anchor_disabled`)
    + Lean `noteSpendV3_grow_gate_forces_set_insert`.

DELETED (the 31-bit debt, greenfield-correct — the proven 8-felt object is now
the only object):
  * circuit: `dsl/revocation.rs` (DslRevocationTree + p3 prove/verify + legacy
    trace), `dsl/tiered_revocation.rs` (rail-built, no consumers),
    `non_revocation_witness.rs`, `non_revocation_adjacency_witness.rs`, the v1
    `dregg-non-revocation-v1` DSL descriptor block, the two `descriptor_by_name`
    goldens + BlindedSet dispatch entries (accumulator rail stays);
  * descriptors: `by-name/non-revocation{,-adjacency}.json` + PROVENANCE rows +
    the emit-script newline-set entry;
  * metatheory: `NonRevocation{,Adjacency}Emit.lean` + the satellite proof
    closure (`NonRevocationRefine{,Complete,Bridge}`, `NonRevocationRung2{,Full}`),
    `EmitNonRevocationAdjacency.lean`, Dregg2.lean/EmitByName.lean registrations
    (the air_accepts-iff-spec schema lives on in its replicas: GarbledEvalRefineBridge,
    EffectVmEmitTransferComplete, EffectVmFullStateRunnableComplete);
  * sdk: the `"non-revocation"` component (prove step 5, verify step-3 arm,
    steps 7/8 host bindings (a)+(b), `expected_revocation_root` params,
    `RevocationRootMismatch`/`NullifierMismatch` variants, `NonRevocationWitness`,
    privacy's `NonRevocationProof`/`prove_not_revoked`/`verify_non_revocation_proof`
    — the polynomial-accumulator rail stays);
  * node: `canonical_revocation_tree/root_for_set` -> `canonical_spent_nullifier_set`
    (capacity 14 -> 65534, the heap tree's 2^16-2), freshness/capability provers
    thread `FullTurnWitness::spent_nullifiers`, bound-verify loses the root arg;
  * consumers: dregg-dsl-runtime re-exports, perf, discord-bot, dregg-tui prose,
    cell doc pointers, 4 circuit-prove rail tests.

KEPT: the SET (now `FullTurnWitness::spent_nullifiers: Option<Vec<BabyBear>>`,
seeding the live 8-felt tree); the sorted-tree sentinels moved to their real
home `heap_root::{SENTINEL_MIN, SENTINEL_MAX}` (cap_root/openable_fields_root
repointed); `revocation_hash_to_field` inlined into the accumulator rail.

REPLACEMENT TEETH (the bindings stay real, not deleted):
  * sdk `freshness_in_circuit_honest_spend_verifies` +
    `freshness_in_circuit_rejects_prover_chosen_set_and_double_spend`
    (prover-chosen set => CommitmentMismatch at 8-felt width; present nullifier
    => prove refuses);
  * node `spend_against_wrong_spent_set_anchor_is_rejected` +
    `spend_freshness_item_cannot_be_substituted` (the declared item cannot dodge
    the in-circuit key).

NOT COMPILED HERE (laptop memory law): the full cargo/tree verify is the
integrator's (persvati/CI) — flagged in the handoff report with the registry
re-emit + limb-37 witness threading.

+++ node/src/api.rs
@@ -323,7 +323,7 @@ pub enum TurnEffectSpec {
         /// Hex cell id. Defaults to the action's target.
         #[serde(default)]
         cell: Option<String>,
-        index: usize,
+        index: u64,
         /// Hex-encoded value: a full 64-char hex field element, or a shorter
         /// hex/decimal scalar that is left-padded into a little-endian u64.
         value: String,
@@ -2492,12 +2492,12 @@ async fn get_receipt_witnesses(
 ///
 /// The proof bytes are persisted by the commit path under
 /// `full_turn_proof:{turn_hash}` (see
-/// [`crate::turn_proving::turn_proof_config_key`]). A spend turn's proof carries
-/// the FRESHNESS-bound non-revocation leg; a light client re-verifying it MUST
-/// pass `expected_revocation_root` = the canonical revocation root derived from
-/// the node's authoritative spent-nullifier set (served via
-/// `nullifier_set_root` on the checkpoint endpoints) into
-/// `dregg_sdk::verify_full_turn_bound`.
+/// [`crate::turn_proving::turn_proof_config_key`]). A spend turn's freshness is
+/// IN-CIRCUIT (the limb-26 grow-gate over the canonical spent set — felt-width
+/// #11 fold-in); a light client re-verifying it MUST pass the CANONICAL
+/// `expected_old_commit` (derived from the authoritative pre-state INCLUDING the
+/// canonical spent-set root) into `dregg_sdk::verify_full_turn_bound` — that
+/// anchor is what binds the opened set to the node's authoritative one.
 ///
 /// Returns the hex-encoded proof bytes plus the turn hash. 404 if no proof was
 /// persisted for that turn (e.g. proving disabled, or the turn carried no
